Make those NSIntegers so some warnings go away

+++ Quicksilver/Code-External/CGSPrivate.h
@@ -53,9 +53,9 @@
 */
 
 // Internal CoreGraphics typedefs
-typedef int		CGSConnection;
-typedef int		CGSWindow;
-typedef int		CGSValue;
+typedef NSInteger	CGSConnection;
+typedef NSInteger   CGSWindow;
+typedef NSInteger   CGSValue;
 
 //// CONSTANTS ////
